ansible/roles/python_jobs_manager/files/src/requirements.txt
requests==2.22.0
oauthlib==3.1.0
pandas==1.2.3
psycopg2==2.8.6
PyYAML==5.4.1
requests_oauthlib==1.3.0
SQLAlchemy==1.4.6
